# internal/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: lowcode-database API
  description: |
    Postgres-backed low-code table service: dynamic tables, columns, rows, indexes,
    relations, and saved queries.

    Optional HTTP webhooks: `/v1/admin/webhooks`. Out of scope (schema bundle import,
    plugins, graph expand, ENUM, RBAC): see docs/roadmap.md.

    **Multi-tenant:** send `X-Tenant-Id` on every request when `TENANT_MODE=multi`.
    In single-tenant mode the header is optional (default tenant from env).

    **Playground:** [lowcode-database-playground](https://github.com/monoposer/lowcode-database-playground) — AG Grid UI (separate repo).

    Errors are returned as plain text (`http.Error`), not JSON envelopes.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local dev (see HTTP_ADDR in .env)
...
paths:
  /v1/admin/database/connection:
    get:
      tags: [Platform]
      summary: Get tenant database connection info (no password)
      operationId: getDatabaseConnection
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDatabaseConnectionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/tenants:
    post:
      tags: [Platform]
      summary: Provision a tenant (multi-tenant mode)
      operationId: createTenant
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTenantResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/types:
    get:
      tags: [Types]
      summary: List built-in column types
      operationId: listTypes
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTypesResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/pg-stat-statements:
    get:
      tags: [Platform]
      summary: List pg_stat_statements for the tenant data database
      description: |
        Requires `PG_STAT_STATEMENTS=true`. When disabled, returns `{ "enabled": false, "statements": [] }`.
        Postgres must load `shared_preload_libraries=pg_stat_statements` and the data migration must create the extension.
      operationId: listPGStatStatements
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPGStatStatementsResponse"
        "400":
          $ref: "#/components/responses/Error"
...
  /v1/admin/schema/er:
    get:
      tags: [Schema]
      summary: Get ER diagram (tables, columns, relations)
      operationId: getERDiagram
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetERDiagramResponse"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: false
      description: |
        Tenant identifier. Required in multi-tenant mode (`TENANT_MODE=multi`) to route
        to the tenant's Postgres database. Also accepted as `X-Tenant-ID`.
      schema:
        type: string
        example: default
...
  responses:
    Error:
      description: Error message (plain text from http.Error)
      content:
        text/plain:
          schema:
            type: string
            example: table_name query parameter is required
  schemas:
...
    Type:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        pgType:
          type: string
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Column:
      type: object
      properties:
        id:
          type: string
        tableName:
          type: string
        name:
          type: string
        typeId:
          type: string
        pgColumn:
          type: string
        isNullable:
          type: boolean
        position:
          type: integer
          format: int32
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateTenantRequest:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        dataDsn:
          type: string
          description: Write DSN (alias of dataDsnWrite)
        dataDsnWrite:
          type: string
        dataDsnReads:
          type: array
          items:
            type: string
          description: Read-replica DSNs (row query / export / saved-query)
        poolMaxConns:
          type: integer
        createDatabase:
          type: boolean
        recordStore:
          type: string
          enum: [shared, dedicated]
          description: >
            shared = public.record on the tenant data DSN (default).
            dedicated = {tenant_id}_record (and matching link_ref / calc_queue) on the same DSN.
            Physical database isolation is dataDsn / dataDsnWrite, not this field.
    CreateTenantResponse:
      type: object
      properties:
        id:
          type: string
        recordStore:
          type: string
          enum: [shared, dedicated]
    ListTypesResponse:
      type: object
      properties:
        types:
          type: array
          items:
            $ref: "#/components/schemas/Type"
    GetDatabaseConnectionResponse:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
          format: int32
        database:
          type: string
        user:
          type: string
        urlWithoutPassword:
          type: string
        psqlCommand:
          type: string
        passwordSourceHint:
          type: string
    ListPGStatStatementsResponse:
      type: object
      properties:
        enabled:
          type: boolean
        statements:
          type: array
          items:
            $ref: "#/components/schemas/PGStatStatement"
    PGStatStatement:
      type: object
      properties:
        queryId:
          type: string
        query:
          type: string
        calls:
          type: integer
          format: int64
        totalExecTimeMs:
          type: number
        meanExecTimeMs:
          type: number
        minExecTimeMs:
          type: number
        maxExecTimeMs:
          type: number
        rows:
          type: integer
          format: int64
        sharedBlksHit:
          type: integer
          format: int64
        sharedBlksRead:
          type: integer
          format: int64
...
    ERNode:
      type: object
      properties:
        tableName:
          type: string
        tableName:
          type: string
        label:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/Column"
    EREdge:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        sourceTableName:
          type: string
        sourceColumnId:
          type: string
        targetTableName:
          type: string
        targetColumnId:
          type: string
        label:
          type: string
    ERDiagram:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ERNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/EREdge"
    GetERDiagramResponse:
      type: object
      properties:
        diagram:
          $ref: "#/components/schemas/ERDiagram"